src/gecko_spec_test/requirements.txt
paramiko~=3.3.1
altwalker~=0.4.0
pybgapi
